refactor: 닉네임 수정 컴포넌트로 리팩토링 #51
    - 유저 페이지의 닉네임 수정 부분을 컴포넌트로 리팩토링 했습니다
    - components 폴더 안에 user 폴더 새로 생성 후 UserNameEdit.tsx 파일 생성
